Fix ImageInterface UI management
