+++ GraphicsView/src/CGAL_Qt5/CMakeLists.txt
@@ -14,8 +14,8 @@ endif()
 
 message( STATUS "USING Qt5_VERSION = '${Qt5Core_VERSION_STRING}'" )
 if(NOT CGAL_HEADER_ONLY)
-  include_directories( ${CMAKE_BINARY_DIR}/Installation/src/CGAL_Qt5 )
   collect_cgal_library( CGAL_Qt5 "${_CGAL_Qt5_MOC_FILES_private};${_CGAL_Qt5_RESOURCE_FILES_private};${_CGAL_Qt5_UI_FILES}")
+  target_include_directories( CGAL_Qt5 PRIVATE ${CMAKE_CURRENT_BINARY_DIR})
 else()
   collect_cgal_library( CGAL_Qt5 "")
 endif()
