+++ Installation/cmake/modules/CGAL_Qt5_moc_and_resource_files.cmake
@@ -15,6 +15,7 @@ if(NOT CGAL_HEADER_ONLY)
     ${CGAL_GRAPHICSVIEW_PACKAGE_DIR}/include/CGAL/Qt/manipulatedCameraFrame.h
     ${CGAL_GRAPHICSVIEW_PACKAGE_DIR}/include/CGAL/Qt/manipulatedFrame.h
     ${CGAL_GRAPHICSVIEW_PACKAGE_DIR}/include/CGAL/Qt/qglviewer.h
+    ${CGAL_GRAPHICSVIEW_PACKAGE_DIR}/include/CGAL/Qt/image_interface.h
     )
 endif()#CGAL_HEADER_ONLY
